backend/requirements.txt
# backend/requirements.txt
# File path: backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
minio==7.2.0
httpx==0.25.2
python-dotenv==1.0.0
bcrypt==4.0.1
openai>=1.0.0
google-genai>=0.2.0